move workspace related things to dedicated files #867
