+++ openeogeotrellis/workspace/__init__.py
@@ -1,0 +1,1 @@
+from .object_storage_workspace import ObjectStorageWorkspace
