+++ openeogeotrellis/workspace/object_storage_workspace.py
@@ -1,7 +1,7 @@
 import logging
 import os
 from pathlib import Path, PurePath
-from typing import Union, Any, Optional
+from typing import Union, Optional
 from urllib.parse import urlparse
 
 import botocore.exceptions
@@ -9,11 +9,11 @@ from boto3.s3.transfer import TransferConfig
 
 from openeo_driver.utils import remove_slash_prefix
 from openeo_driver.workspace import Workspace, _merge_collection_metadata
-from pystac import STACObject, Collection, CatalogType, Link, Item, Asset
+from pystac import STACObject, Collection, CatalogType, Item, Asset
 from pystac.layout import HrefLayoutStrategy, CustomLayoutStrategy
-from pystac.stac_io import DefaultStacIO
 
 from openeogeotrellis.utils import s3_client
+from .custom_stac_io import CustomStacIO
 
 _log = logging.getLogger(__name__)
 
@@ -169,33 +169,3 @@ class ObjectStorageWorkspace(Workspace):
             raise ValueError(asset_uri)
 
         return f"s3://{self.bucket}/{target_key}"
-
-
-# TODO: move to dedicated file?
-class CustomStacIO(DefaultStacIO):
-    """Adds support for object storage."""
-
-    @property
-    def _s3(self):
-        # otherwise there's an infinite recursion error upon Item.get_assets() wrt/ some boto3 reference
-        return s3_client()
-
-    def read_text(self, source: Union[str, Link], *args: Any, **kwargs: Any) -> str:
-        parsed = urlparse(source)
-        if parsed.scheme == "s3":
-            bucket = parsed.netloc
-            key = parsed.path[1:]
-
-            obj = self._s3.get_object(Bucket=bucket, Key=key)
-            return obj["Body"].read().decode("utf-8")
-        else:
-            return super().read_text(source, *args, **kwargs)
-
-    def write_text(self, dest: Union[str, Link], txt: str, *args: Any, **kwargs: Any) -> None:
-        parsed = urlparse(dest)
-        if parsed.scheme == "s3":
-            bucket = parsed.netloc
-            key = parsed.path[1:]
-            self._s3.put_object(Bucket=bucket, Key=key, Body=txt.encode("utf-8"), ContentEncoding="utf-8")
-        else:
-            super().write_text(dest, txt, *args, **kwargs)
